Reset Password Styling Done
