+++ src/App.js
@@ -1,7 +1,5 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import { useSearchParams } from "react-router-dom";
-import { useLocation } from "react-router-dom"
 import "./App.css";
 
 // Pages
@@ -9,18 +7,20 @@ import Canvas from "./Pages/canvas.js";
 import Settings from "./Pages/settings.js";
 import PrivacyPolicy from "./Pages/privacypol.js";
 import TermsAndConditions from "./Pages/terms&con.js";
-import ForgotPassword from "./Pages/ForgotPassword";
+import ResetPassword from "./Pages/resetPass";
 
 function App() {
-
   return (
     <Router>
       <Routes>
-        <Route  path="/" element = {<Canvas />}></Route>
-        <Route  path="Settings" element = {<Settings />}></Route>
-        <Route  path="PrivacyPolicy" element={<PrivacyPolicy />}></Route>
-        <Route  path="TermsAndConditions" element = {<TermsAndConditions />}></Route>
-        <Route  path="ForgotPassword/*" element = {<ForgotPassword />}></Route>
+        <Route path="/" element={<Canvas />}></Route>
+        <Route path="Settings" element={<Settings />}></Route>
+        <Route path="PrivacyPolicy" element={<PrivacyPolicy />}></Route>
+        <Route
+          path="TermsAndConditions"
+          element={<TermsAndConditions />}
+        ></Route>
+        <Route path="ResetPassword/*" element={<ResetPassword />}></Route>
       </Routes>
     </Router>
   );
