+++ src/Components/Styling/signupModal.css
@@ -39,6 +39,7 @@
 
 .signupInput {
   position: relative;
+  display: flex;
   width: 70%;
   margin-top: 35px;
   margin-bottom: 0px;
@@ -65,6 +66,7 @@
   font-weight: normal;
   position: absolute;
   padding: 15px 10px;
+  bottom: 0px;
   pointer-events: none;
   transition: 0.5s;
   /* outline: 1px solid red; */
